move eval to separate package
